Introduce new header files for SoC register defines
As it turned out, most of QC/A WiSoC are very similar
from register structures and addresses point of view.

What's more, it seems that even when some registers
or fields are slightly different in name, they still
mean the same and work in exactly same way.

So, instead of using approach from old Atheros SDK,
where we have hundreds of different header files,
for different chips, we will use only one. And in
case of different address for the same register or
different fields size/position, etc., we will use
preprocessor and SOC_TYPE value.

Add also MediaTek/Ralink headers file, we are going
to fill it soon.