Allow loading kernel from rootfs+kernel and kernel+rootfs firmware images
Most of the boards with firmware based on official and old QC/A SDK use
rootfs+kernel firmware order. In OpenWrt/LEDE kernel+rootfs order is
prefered (better for growing size of kernel) and more common.

As we don't know what firmware user is using, make it working with both
type of images, with kernel in front or behind the rootfs.

With this change, there is no need to change 'bootcmd' variable when
switching from vendor firmware to OpenWrt/LEDE on YunCore AP90Q/CPE830.