Add missing ART offset for supported COMFAST devices
In most devices, last sector is used for storing ART
data. It's true also for COMFAST devices, as it looks
that they have two copies of ART data, inside second
and last FLASH sector (named 'nvram' in OFW).

ART in second sector is (will be) used in OpenWrt/LEDE,
so we will also use it as default one.