Do not use REF clock divider higher than 2 on QCA95xx/AR934x
As tests showed, REF clock divider is not reliable with
values higher than 2. Final clock frequency is stable
between restarts, but its value is in range +/- 20-25%
in comparison to expected/calculated frequency.

Instead of using higher REF clock divider values, make
use of fractional part of the multiplier.

This fixes #128