Add support for multikey autoboot stop string
This allows to define string (in environment variable 'bootstopkey')
used to interrupt autobooting process. It works in similar way as in
upstream U-Boot version.

If variable is not defined in environment, default approach (any key)
will be used. Variable can be also predefined during compilation with
CONFIG_AUTOBOOT_STOP_STR define.

This closes #136