In case of missing recovery script in env, copy it from default one
In case of update from some old version on device with correct env in
FLASH, recovery script (moved to env variable 'recovery' in 4f11127)
isn't available which results in missing 'button recovery mode'.

This fixes backward compatibility and ensures that the 'recovery'
variable is available in env - in case if it's not available in saved
env, value is copied from default env (the one embedded with image).

This fixes #154 and closes #131