limit USB fix to QCA953X
USB fix rendered USB port unusable on QCA9558. Not sure yet if chipset
difference is the cause or USB wiring design (GPIO controllable USB power
on Archer C7). At least the original issue this fix was created for is not
present with QCA u-boot on USB1 but still it is present on USB2 port.

Signed-off-by: Tomislav Požega <pozega.tomislav@gmail.com>